packages/sre-agent/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sre-agent"
version = "0.1.0"
description = "Python SRE incident-diagnosis agent for the monorepo."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "chromadb>=1.5,<2.0",
  "langchain-chroma>=1.1,<2.0",
  "langchain-community>=0.4,<1.0",
  "langchain-openai>=1.1,<2.0",
  "langchain>=1.2,<2.0",
  "python-dotenv>=1.0,<2.0"
]

[project.scripts]
sre-agent-ingest = "sre_agent.ingest:main"
sre-agent-query = "sre_agent.query:main"
sre-agent-diagnose = "sre_agent.agent:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
